requirements.txt
dj-database-url==0.4.0
Django==1.9.4
django-filter==0.13.0
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
gunicorn==19.4.5
Markdown==2.6.6
whitenoise==2.0.6
psycopg2==2.6.1
django-rest-swagger==0.3.5
django-cors-headers==1.1.0